* Use the buildts and git commit in versioning / upgrading
Now using buildts again for determining if a bot should be upgraded.
Buildts is the UTC timestamp of the last change committed to a branch.
Now also appending the branch to all versions displayed in the bot.
Also showing the short commit in place of the old svn revision.
Updates offer svn rev 9999 for older bots.