* Fix 'HQ' user getting wiped in -t mode resulting in a crash. (fixes #298)
The user is no longer booted after the user transfer. Instead, if it did
not exist on the remote side, it is re-added. This resolves the segfault
from not having a user entry.

This may seem to be a security risk, however, adding user 'HQ' with minimal
flags will mitigate this a little because then the linking bot will still
have its flags wiped, but retain a user. This is of course assuming the binary
has not been tampered with.