* Add -Q to read PackConfig securely over stdin
This will disable echoing if possible so that snooping on
the TTY will not reveal the PackConfig. This will remove
the need to upload the pack.cfg to servers to initialize the
binary. This all assumes safe libraries/kernel/terminal.

The reading halts when it sees SALT2. Failure to provide SALT2
will make it appear to hang. Hitting ^D will exit out of the
session. Also, it may be necessary to hit <enter> at the
end of the paste.