* Don't fail a server if it sends no answers
This could simply be that there are no answers... query A or AAAA and receive nothing back...
With the new NXDOMAIN / SERVFAIL error checks, this hard error of jumping servers is probably
no longer needed, as there is always a good reason for not getting answers back.

If the server has a permanent error, it lets us know now.