* Fix glibc 2.10.1 issues
glibc 2.10.1 added a 2nd strchr()/strrchr() prototype, resulting in:
char *strchr(char *, int)
const char *strchr(const char *, int)

This causes code passing in a (const char*) string to return a
(const char*) result. This causes (invalid) code to be revealed.

This patch fixes the compile problem, but not the broken code. This
code is likely destroying valid string buffers and will be fixed in
1.2.17.