* Don't enforce_bitch() on the trigger.
+bitch will trigger, punishing the opper and anyone opped. reversed
is set, so anyone deopped will get reopped. The botnet will
send +botbitch, causing the channel to be rechecked and thus enforce_bitch
called very shortly. Better to act on what just happened, instead
of mass deopping in a random manor, as the opped clients may be last
to get deopped in the enforce_bitch().