Add new consistent dynamic roles
This will automatically assign roles to every bot in a channel
in a decentralized way. This ensures that both sides of a split
will have a bot covering every role and removes the need for
the leaf to share nicks to the hub and for the hub to have
to process roles continually.