Attempt to use PR_SET_PTRACER even if not defined at build-time to fix foward-compat.
A binary built on Linux 2.6 will not have PR_SET_PTRACER defined, but
when ran on Linux 3.4+ will still hit the ptrace limitations if
/proc/sys/kernel/yama/ptrace_scope is set to 1.  Work around this by
hardcoding the PR_SET_PTRACER define and trying it anyway.  The result
on older kernel is just an [EINVAL].  This is largely redundant with the
PR_SET_DUMPABLE change but is still worth keeping.