Must share SET entries to all bots if the target is not online.
The problem is if a new bot is added and they are not yet connected,
then a new .set is ran.  The new SET entry will never share to the
localhub.  Once the bot does connect it comes up with stale information
since the localhub is caching the userfile directly to it.  This is not
resolved until the localhub relinks to get a fresh userfile AND their
leaf bots are connected.

This reworks 1da96702 and 174b3caa.